Ajout de Sécurité : Register et login. - beans Register, login - Interface et Dao - Interface et Service - Couche présentaion - page jsp register et login.

+++ src/main/java/com/mamouz/cv/dao/interfaces/ITitreCV.java
@@ -3,6 +3,6 @@ package com.mamouz.cv.dao.interfaces;
 import com.mamouz.cv.entity.TitreCV;
 import com.mamouz.cv.service.interfaces.IService;
 
-public interface ITitreCV extends IService<TitreCV, Integer> {
+public interface ITitreCV extends IService<TitreCV, Long> {
 
 }
