+++ src/main/java/com/mamouz/cv/service/interfaces/ITitreService.java
@@ -9,6 +9,6 @@ import com.mamouz.cv.entity.TitreCV;
  * @author mikak
  *
  */
-public interface ITitreService extends IService<TitreCV, Integer> {
+public interface ITitreService extends IService<TitreCV, Long> {
 
 }
